memperbaiki dashboard petugas

+++ routes/web.php
@@ -39,7 +39,7 @@ Route::get('konfirmasi-pinjaman', [\App\Http\Controllers\PinjamController::class
 // URL for Search Barang
 Route::get('/search', [\App\Http\Controllers\PageUserController::class, 'search'])->name('search')->middleware('auth');
 
-// URL for Petugas Peralatan 
+// URL for Petugas Peralatan and give middleware to group
 Route::prefix('petugas_peralatan')->group(function () {
     Route::get('/login-petugas', [App\Http\Controllers\Auth\PetugasPeralatanLoginController::class, 'showLoginForm'])->name('petugas_peralatan.login');
     Route::post('/login-petugas', [App\Http\Controllers\Auth\PetugasPeralatanLoginController::class, 'login'])->name('petugas_peralatan.login.submit');
@@ -48,21 +48,21 @@ Route::prefix('petugas_peralatan')->group(function () {
     Route::get('/dashboard', [App\Http\Controllers\PetugasPeralatanController::class, 'index'])->name('petugas_peralatan.dashboard');
 
     // URL for Mahasiswa & Dosen
-    Route::get('/mahasiswa', [App\Http\Controllers\PageUserController::class, 'getUser'])->name('petugas_peralatan.mahasiswa');
-    Route::get('/dosen', [App\Http\Controllers\PageUserController::class, 'getDosen'])->name('petugas_peralatan.dosen');
+    Route::get('/mahasiswa', [App\Http\Controllers\PageUserController::class, 'getUser'])->name('petugas_peralatan.mahasiswa')->middleware('auth:petugas_peralatan');
+    Route::get('/dosen', [App\Http\Controllers\PageUserController::class, 'getDosen'])->name('petugas_peralatan.dosen')->middleware('auth:petugas_peralatan');
 
     // URL for Barang
-    Route::get('/barang', [App\Http\Controllers\BarangController::class, 'index'])->name('barang.index');
-    Route::get('/barang/create', [App\Http\Controllers\BarangController::class, 'create'])->name('barang.create');
-    Route::post('/barang', [App\Http\Controllers\BarangController::class, 'store'])->name('barang.store');
-    Route::get('/barang/{id}', [App\Http\Controllers\BarangController::class, 'show'])->name('barang.show');
-    Route::get('/barang/{id}/edit', [App\Http\Controllers\BarangController::class, 'edit'])->name('barang.edit');
-    Route::put('/barang/{id}', [App\Http\Controllers\BarangController::class, 'update'])->name('barang.update');
-    Route::delete('/barang/{id}', [App\Http\Controllers\BarangController::class, 'destroy'])->name('barang.destroy');
-    
-    Route::get('pengembalian-barang', [App\Http\Controllers\PengembalianController::class, 'pengembalian'])->name('petugas_peralatan.pengembalian');
-    Route::get('pengembalian-detail/{id}', [App\Http\Controllers\PengembalianController::class, 'detailPengembalian'])->name('petugas_peralatan.detailPengembalian');
-    Route::post('pengembalian-barang/{id}/{id_barang}', [App\Http\Controllers\PengembalianController::class, 'pengembalianBarang'])->name('petugas_peralatan.pengembalianBarang');
+    Route::get('/barang', [App\Http\Controllers\BarangController::class, 'index'])->name('barang.index')->middleware('auth:petugas_peralatan');
+    Route::get('/barang/create', [App\Http\Controllers\BarangController::class, 'create'])->name('barang.create')->middleware('auth:petugas_peralatan');
+    Route::post('/barang', [App\Http\Controllers\BarangController::class, 'store'])->name('barang.store')->middleware('auth:petugas_peralatan');
+    Route::get('/barang/{id}', [App\Http\Controllers\BarangController::class, 'show'])->name('barang.show')->middleware('auth:petugas_peralatan');
+    Route::get('/barang/{id}/edit', [App\Http\Controllers\BarangController::class, 'edit'])->name('barang.edit')->middleware('auth:petugas_peralatan');
+    Route::put('/barang/{id}', [App\Http\Controllers\BarangController::class, 'update'])->name('barang.update')->middleware('auth:petugas_peralatan');
+    Route::delete('/barang/{id}', [App\Http\Controllers\BarangController::class, 'destroy'])->name('barang.destroy')->middleware('auth:petugas_peralatan');
+
+    Route::get('pengembalian-barang', [App\Http\Controllers\PengembalianController::class, 'pengembalian'])->name('petugas_peralatan.pengembalian')->middleware('auth:petugas_peralatan');
+    Route::get('pengembalian-detail/{id}', [App\Http\Controllers\PengembalianController::class, 'detailPengembalian'])->name('petugas_peralatan.detailPengembalian')->middleware('auth:petugas_peralatan');
+    Route::post('pengembalian-barang/{id}/{id_barang}', [App\Http\Controllers\PengembalianController::class, 'pengembalianBarang'])->name('petugas_peralatan.pengembalianBarang')->middleware('auth:petugas_peralatan');
 });
